created weather forecast card component
